observability/local/log-sentinel/requirements.txt
flask>=3.0.0
requests>=2.31.0
pyyaml>=6.0.1
schedule>=1.2.0
pytest>=8.0.0
